# APItest/SwaggerUI api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SK API",
    "description": "SK智慧校园API文档",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "v114.514"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/updatePwd/{oldPwd}/{newPwd}": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "修改密码",
        "description": "修改密码接口",
        "operationId": "updatePwd",
        "parameters": [
          {
            "name": "oldPwd",
            "in": "path",
            "description": "旧密码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newPwd",
            "in": "path",
            "description": "新密码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/studentController/addOrUpdateStudent": {
      "post": {
        "tags": [
          "Student"
        ],
        "summary": "保存或者修改学生信息",
        "description": "保存或者修改学生信息",
        "operationId": "addOrUpdateStudent",
        "parameters": [
          {
            "name": "student",
            "in": "query",
            "description": "要保存或修改的学生JSON",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "登录",
        "description": "登录接口",
        "operationId": "login",
        "parameters": [
          {
            "name": "loginForm",
            "in": "query",
            "description": "登录表单",
            "required": true
          },
          {
            "name": "request",
            "in": "query",
            "description": "请求",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/headerImgUpload": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "上传图片",
        "description": "上传图片接口",
        "operationId": "headerImgUpload",
        "parameters": [
          {
            "name": "multipartFile",
            "in": "query",
            "description": "图片文件",
            "required": true,
            "schema": {
              "type": "string",
              "format": "binary"
            }
          },
          {
            "name": "request",
            "in": "query",
            "description": "请求",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/studentController/getStudentByOpr/{pageNo}/{pageSize}": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "分页带条件查询学生信息",
        "description": "分页带条件查询学生信息",
        "operationId": "getStudentBuOpr",
        "parameters": [
          {
            "name": "pageNo",
            "in": "path",
            "description": "页码数",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "path",
            "description": "页大小",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "student",
            "in": "query",
            "description": "查询的条件",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Student"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/getVerifiCodeImage": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "获取验证码图片",
        "description": "获取验证码图片接口",
        "operationId": "getVerifiCodeImage",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "description": "请求",
            "required": true
          },
          {
            "name": "response",
            "in": "query",
            "description": "响应",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/getInfo": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "获取用户信息",
        "description": "通过token口令获取当前登录的用户信息",
        "operationId": "getInfoByToken",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/studentController/delStudentById": {
      "delete": {
        "tags": [
          "Student"
        ],
        "summary": "删除学生",
        "description": "删除单个或者多个学生信息",
        "operationId": "delStudentById",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "要删除的学生编号的JSON数组",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResultObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "数据",
            "example": null
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "学生Id",
            "format": "int32",
            "example": 1
          },
          "sno": {
            "type": "string",
            "description": "学生学号",
            "example": "2018111111"
          },
          "name": {
            "type": "string",
            "description": "学生姓名",
            "example": "张三"
          },
          "gender": {
            "type": "string",
            "description": "学生性别",
            "example": "男"
          },
          "password": {
            "type": "string",
            "description": "学生密码",
            "example": "123456"
          },
          "email": {
            "type": "string",
            "description": "学生邮箱",
            "example": "2421361985"
          },
          "telephone": {
            "type": "string",
            "description": "学生电话",
            "example": "18888888888"
          },
          "address": {
            "type": "string",
            "description": "学生地址",
            "example": "广东省广州市"
          },
          "introducation": {
            "type": "string",
            "description": "学生介绍",
            "example": "我是一个学生"
          },
          "portrait": {
            "type": "string",
            "description": "学生头像",
            "example": "http://localhost:8080/sms/studentController/getPortrait/1"
          },
          "portraitPath": {
            "type": "string",
            "description": "学生头像路径",
            "example": "http://localhost:8080/sms/studentController/getPortrait/1"
          }
        }
      },
...
      "LoginForm": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "123456"
          },
          "verifiCode": {
            "type": "string",
            "description": "验证码",
            "example": "1234"
          },
          "userType": {
            "type": "integer",
            "description": "用户类型",
            "format": "int32",
            "example": 1
          }
        }
      }
    }
  }
}